Separate out no_mangle functions, remove pure Rust stuff from c_api module.
