+++ crates/xetex_layout/src/lib.rs
@@ -8,6 +8,22 @@
 //!
 //! [Tectonic]: https://tectonic-typesetting.github.io/
 
+macro_rules! cstr {
+    ($lit:literal) => {
+        // SAFETY: C string passed to from_ptr guaranteed to end with a null due to concat!
+        unsafe {
+            ::std::ffi::CStr::from_ptr(
+                ::std::ptr::from_ref(concat!($lit, "\0")).cast::<::libc::c_char>(),
+            )
+        }
+    };
+}
+
+mod engine;
+mod font;
+mod manager;
+mod utils;
+
 /// Import things from our bridge crates to ensure that we actually link with
 /// them.
 mod linkage {
@@ -31,22 +47,10 @@ mod linkage {
     use tectonic_bridge_fontconfig as clipyrenamehack6;
 }
 
-macro_rules! cstr {
-    ($lit:literal) => {
-        // SAFETY: C string passed to from_ptr guaranteed to end with a null due to concat!
-        unsafe {
-            ::std::ffi::CStr::from_ptr(
-                ::std::ptr::from_ref(concat!($lit, "\0")).cast::<::libc::c_char>(),
-            )
-        }
-    };
-}
-
-mod c_api {
-    use crate::c_api::engine::LayoutEngine;
-    use crate::c_api::font::Font;
+pub(crate) mod c_api {
+    use crate::engine::LayoutEngine;
+    use crate::font::Font;
     use std::collections::BTreeMap;
-    use std::ptr::NonNull;
     use std::sync::Mutex;
     #[cfg(not(target_os = "macos"))]
     use tectonic_bridge_fontconfig as fc;
@@ -58,18 +62,18 @@ mod c_api {
     #[derive(Copy, Clone, PartialEq, Debug)]
     #[repr(C)]
     pub struct FloatPoint {
-        x: f32,
-        y: f32,
+        pub x: f32,
+        pub y: f32,
     }
 
     /// cbindgen:rename-all=camelCase
     #[derive(Copy, Clone, Default, PartialEq, Debug)]
     #[repr(C)]
     pub struct GlyphBBox {
-        x_min: f32,
-        y_min: f32,
-        x_max: f32,
-        y_max: f32,
+        pub x_min: f32,
+        pub y_min: f32,
+        pub x_max: f32,
+        pub y_max: f32,
     }
 
     #[cfg(not(target_os = "macos"))]
@@ -84,34 +88,15 @@ mod c_api {
     /// cbindgen:ignore
     pub type XeTeXLayoutEngine = *mut LayoutEngine;
     #[cfg(not(target_os = "macos"))]
-    type RawPlatformFontRef = *mut fc::sys::FcPattern;
+    pub(crate) type RawPlatformFontRef = *mut fc::sys::FcPattern;
     #[cfg(target_os = "macos")]
-    type RawPlatformFontRef = tectonic_mac_core::sys::CTFontDescriptorRef;
+    pub(crate) type RawPlatformFontRef = tectonic_mac_core::sys::CTFontDescriptorRef;
     /// cbindgen:ignore
     #[cfg(not(target_os = "macos"))]
-    type PlatformFontRef = fc::Pattern;
+    pub(crate) type PlatformFontRef = fc::Pattern;
     /// cbindgen:ignore
     #[cfg(target_os = "macos")]
-    type PlatformFontRef = tectonic_mac_core::CTFontDescriptor;
-
-    fn fix_to_d(f: Fixed) -> f64 {
-        f as f64 / 65536.0
-    }
-
-    fn d_to_fix(d: f64) -> Fixed {
-        (d * 65536.0 + 0.5) as Fixed
-    }
-
-    fn raw_to_rs(font: RawPlatformFontRef) -> Option<PlatformFontRef> {
-        #[cfg(target_os = "macos")]
-        let out = {
-            use tectonic_mac_core::CoreType;
-            NonNull::new(font.cast_mut()).map(PlatformFontRef::new_borrowed)
-        };
-        #[cfg(not(target_os = "macos"))]
-        let out = { unsafe { NonNull::new(font).map(|p| PlatformFontRef::from_raw_borrowed(p)) } };
-        out
-    }
+    pub(crate) type PlatformFontRef = tectonic_mac_core::CTFontDescriptor;
 
     /// key is combined value representing `(font_id << 16) + glyph`
     /// value is glyph bounding box in TeX points
